Innfører AvrundetBeløp for å representere et heltallsbeløp

+++ domene/src/main/kotlin/no/nav/aap/domene/utbetaling/modellapi/Paragraf_11_19_3_leddModellApi.kt
@@ -5,6 +5,6 @@ import java.time.LocalDate
 data class Paragraf_11_19_3_leddModellApi(
     val dato: LocalDate,
     val grunnlagsfaktor: Double,
-    val grunnbeløp: Double,
+    val grunnbeløp: Int,
     val grunnlag: Double
 )
